set listener on runtime and invoke per-query listeners
