+++ ksqldb-engine/src/main/java/io/confluent/ksql/util/PersistentQueryMetadataImpl.java
@@ -31,7 +31,6 @@ import io.confluent.ksql.name.SourceName;
 import io.confluent.ksql.physical.scalablepush.ScalablePushRegistry;
 import io.confluent.ksql.query.KafkaStreamsBuilder;
 import io.confluent.ksql.query.MaterializationProviderBuilderFactory;
-import io.confluent.ksql.query.QueryError;
 import io.confluent.ksql.query.QueryErrorClassifier;
 import io.confluent.ksql.query.QueryId;
 import io.confluent.ksql.schema.ksql.PhysicalSchema;
@@ -41,7 +40,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
-import org.apache.kafka.streams.KafkaStreams.State;
 import org.apache.kafka.streams.Topology;
 import org.apache.kafka.streams.errors.StreamsUncaughtExceptionHandler;
 
@@ -233,32 +231,4 @@ public class PersistentQueryMetadataImpl
     return scalablePushRegistry;
   }
 
-  private static final class QueryListenerWrapper implements Listener {
-    private final Listener listener;
-    private final Optional<ScalablePushRegistry> scalablePushRegistry;
-
-    private QueryListenerWrapper(final Listener listener,
-        final Optional<ScalablePushRegistry> scalablePushRegistry) {
-      this.listener = listener;
-      this.scalablePushRegistry = scalablePushRegistry;
-    }
-
-    @Override
-    public void onError(final QueryMetadata queryMetadata, final QueryError error) {
-      this.listener.onError(queryMetadata, error);
-      scalablePushRegistry.ifPresent(ScalablePushRegistry::onError);
-    }
-
-    @Override
-    public void onStateChange(final QueryMetadata queryMetadata, final State before,
-        final State after) {
-      this.listener.onStateChange(queryMetadata, before, after);
-    }
-
-    @Override
-    public void onClose(final QueryMetadata queryMetadata) {
-      this.listener.onClose(queryMetadata);
-      scalablePushRegistry.ifPresent(ScalablePushRegistry::cleanup);
-    }
-  }
 }
